chore: move TTS stress test to lib/glific/scripts for release availability
Compiled into the release so it can be called directly from
the remote IEx console without pasting code.
